implement exception handler in lab3
1.implement exception handler in asm
2.save context when exception happen

+++ lab3/include/string.h
@@ -7,7 +7,7 @@ char *strcpy(char *dest, char *src);
 int strlen(const char *str);
 void strset(char *dest, int val, int size);
 void itoa(int x, char *str);
-void int2hex(unsigned int val, char *str);
+void int2hex(unsigned long int val, char *str);
 void ftoa(float val, char *str);
 
 #endif
